requirements.txt
Flask==2.3.3
Flask-Assets==2.0
Flask-Babel==3.1.0
Babel==2.12.1
cssmin==0.2.0
jsmin==3.0.1
python-telegram-bot==22.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
gunicorn==21.2.0 